+++ R/show_ex_toc.R
@@ -5,12 +5,15 @@
 #' [learnr()], [assignment()], and [challenge()] respectively. This function
 #' creates the table of content for all these exercises.
 #'
+#' @param id An identifier for this table of content.
 #' @param header A Markdown text to place as header of the exercises toc.
 #' @param clear.it Do we clear the toc list (`TRUE` by default)?
+#' @param finalize If `TRUE`, we also compile the `assignments.csv` and
+#' `apps.csv` files with all assignments and all apps.
 #'
 #' @return The Markdown chunk with the exercises toc.
 #' @export
-show_ex_toc <- function(header = "", clear.it = TRUE) {
+show_ex_toc <- function(id, header = "", clear.it = TRUE, finalize = FALSE) {
   ex_toc <- getOption("learnitdown_ex_toc", NULL)
   if (is.null(ex_toc)) {
     toc <- ""
@@ -20,20 +23,32 @@ show_ex_toc <- function(header = "", clear.it = TRUE) {
       options(learnitdown_ex_toc = NULL)
   }
 
-  # Compile a single assignments.csv file
+  # Compile an apps_id.csv file per module,
   ex_dir <- file.path(.get_output_dir(), "ex")
-  files <- dir(ex_dir, pattern = "^assignment_.+\\.csv$", full.names = TRUE)
-  res <- data.frame()
-  for (file in files)
-    res <- rbind(res, read.csv(file))
-  write.csv(res, file.path(ex_dir, "assignments.csv"), row.names = FALSE)
-
-  # Also compile a apps.csv file
   apps <- getOption("learnitdown_apps", data.frame())
-  apps <- rbind(apps, res)
-  # Sort by app name
-  apps <- apps[order(apps$app), ]
-  write.csv(apps, file.path(ex_dir, "apps.csv"), row.names = FALSE)
+  write.csv(apps, file.path(ex_dir, paste0("apps_", id, ".csv")),
+    row.names = FALSE)
+  if (isTRUE(clear.it))
+    options(learnitdown_apps = NULL)
+
+  if (isTRUE(finalize)) {
+    # Compile a single assignments.csv file
+    files <- dir(ex_dir, pattern = "^assignment_.+\\.csv$", full.names = TRUE)
+    res <- data.frame()
+    for (file in files)
+      res <- rbind(res, read.csv(file))
+    write.csv(res, file.path(ex_dir, "assignments.csv"), row.names = FALSE)
+
+    # Also compile an apps.csv with all apps and assignments
+    apps_files <- dir(ex_dir, pattern = "^apps_.+\\.csv$", full.names = TRUE)
+    apps_res <- data.frame()
+    for (app_file in apps_files)
+      apps_res <- rbind(apps_res, read.csv(app_file))
+    all_apps <- rbind(apps_res, res)
+    # Sort by app name
+    all_apps <- all_apps[order(all_apps$app), ]
+    write.csv(all_apps, file.path(ex_dir, "apps.csv"), row.names = FALSE)
+  }
 
   toc
 }